# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Digital Marketplace API",
        "description": "API documentation for Digital Marketplace - A platform for creators to sell digital products",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Local API Server"
        }
    ],
    "paths": {
...
        "/api/v1/checkout/{product}": {
            "post": {
                "tags": [
                    "Purchases"
                ],
                "summary": "Checkout a product",
                "description": "Purchase a product (simulated payment). Customer only. Cannot purchase same product twice.",
                "operationId": "66ee1d8f5d244b0d3ae03056a57b22ce",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Product purchased successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CheckoutResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthenticated"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - Customer role required",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Forbidden"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Business logic error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "You have already purchased this product."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFound"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/library": {
            "get": {
                "tags": [
                    "Purchases"
                ],
                "summary": "Get user's library",
                "description": "Retrieve a paginated list of purchased products (library). Customer only.",
                "operationId": "958e1712bab9232b8cb2072192c0c7ed",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15,
                            "example": 15
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PurchasePaginated"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthenticated"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - Customer role required",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Forbidden"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/library/{purchase}": {
            "get": {
                "tags": [
                    "Purchases"
                ],
                "summary": "Get specific purchase details",
                "description": "Retrieve details of a specific purchase including download statistics. Customer only.",
                "operationId": "dacfa711b8021d9ef62adc2e6455ba1a",
                "parameters": [
                    {
                        "name": "purchase",
                        "in": "path",
                        "description": "Purchase ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Purchase"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthenticated"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - Not your purchase or Customer role required",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "You do not have access to this purchase."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Purchase not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFound"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/sales": {
            "get": {
                "tags": [
                    "Purchases"
                ],
                "summary": "Get creator's sales",
                "description": "Retrieve a paginated list of sales for products created by the authenticated creator. Creator only.",
                "operationId": "ac034daded923414ee8f56193ce3ced0",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15,
                            "example": 15
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PurchasePaginated"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthenticated"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - Creator role required",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Forbidden"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/revenue": {
            "get": {
                "tags": [
                    "Purchases"
                ],
                "summary": "Get creator's revenue statistics",
                "description": "Retrieve revenue statistics for products created by the authenticated creator. Creator only.",
                "operationId": "83ff8cc1e2862508c017139578f1f997",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CreatorRevenue"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthenticated"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - Creator role required",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Forbidden"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Unauthenticated": {
                "title": "Unauthenticated",
                "description": "Unauthenticated response",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unauthenticated."
                    }
                },
                "type": "object"
            },
            "Forbidden": {
                "title": "Forbidden",
                "description": "Forbidden response",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "This action is unauthorized."
                    }
                },
                "type": "object"
            },
            "NotFound": {
                "title": "Not Found",
                "description": "Resource not found response",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Resource not found"
                    }
                },
                "type": "object"
            },
...
            "Purchase": {
                "title": "Purchase",
                "description": "Purchase model",
                "required": [
                    "id",
                    "user_id",
                    "product_id",
                    "reference",
                    "status"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 10
                    },
                    "product_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "reference": {
                        "type": "string",
                        "example": "PUR-ABC123DEF456"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "completed",
                            "failed"
                        ],
                        "example": "completed"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-01-27T03:16:30.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-01-27T03:16:30.000000Z"
                    },
                    "product": {
                        "$ref": "#/components/schemas/Product"
                    },
                    "user": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/User"
                            }
                        ],
                        "nullable": true,
                        "type": "object"
                    },
                    "download_logs": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/DownloadLog"
                        },
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "PurchasePaginated": {
                "title": "Paginated Purchases",
                "description": "Paginated list of purchases",
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Purchase"
                        }
                    },
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "first_page_url": {
                        "type": "string",
                        "example": "http://localhost:8000/api/v1/library?page=1"
                    },
                    "from": {
                        "type": "integer",
                        "example": 1
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 5
                    },
                    "last_page_url": {
                        "type": "string",
                        "example": "http://localhost:8000/api/v1/library?page=5"
                    },
                    "links": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "url": {
                                    "type": "string",
                                    "nullable": true
                                },
                                "label": {
                                    "type": "string"
                                },
                                "active": {
                                    "type": "boolean"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "next_page_url": {
                        "type": "string",
                        "example": "http://localhost:8000/api/v1/library?page=2",
                        "nullable": true
                    },
                    "path": {
                        "type": "string",
                        "example": "http://localhost:8000/api/v1/library"
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 15
                    },
                    "prev_page_url": {
                        "type": "string",
                        "nullable": true
                    },
                    "to": {
                        "type": "integer",
                        "example": 15
                    },
                    "total": {
                        "type": "integer",
                        "example": 50
                    }
                },
                "type": "object"
            },
            "CheckoutResponse": {
                "title": "Checkout Response",
                "description": "Response after successful checkout",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Product purchased successfully"
                    },
                    "purchase": {
                        "$ref": "#/components/schemas/Purchase"
                    }
                },
                "type": "object"
            },
...
            "CreatorRevenue": {
                "title": "Creator Revenue Statistics",
                "description": "Revenue statistics for creator",
                "properties": {
                    "total_sales": {
                        "type": "integer",
                        "example": 150
                    },
                    "total_revenue": {
                        "type": "string",
                        "example": "7499.50"
                    },
                    "average_order_value": {
                        "type": "string",
                        "example": "49.99"
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "User",
                "description": "User model",
                "required": [
                    "id",
                    "name",
                    "email",
                    "role"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john@example.com"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "creator",
                            "customer"
                        ],
                        "example": "creator"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-01-27T03:16:30.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-01-27T03:16:30.000000Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}